Extend user profile, add possibility to track habits progress for today.
Add picture, locale and time zone to the profile, give the possibility to change the time zone so the user sees the history dates in selected time zone and formatted and translated according to her locale.

Refactor services according to functionality to separate profile, habits and the history which increases readability and makes it easier to change / add some functionality in future.
Add history endpoints. Store habits history to allow user to save the progress on each habit and review it later.

Preload data in App component to fix the continuous reload problem.