Refactoring (#3)
Update colors so the site looks at least a bit better.

Extend user profile. Add picture for more personalized experience. Add locale and time zone to the profile as first step of internationalization. Give the possibility to change the time zone so the user sees the history dates in selected time zone and formatted and translated according to her locale.

Refactor services according to functionality to separate profile, habits and the history which increases readability and makes it easier to change / add some functionality in future.
Add history endpoints. Store habits history to allow user to save the progress on each habit and review it later.

Preload data in App component to fix the continuous reload problem.

Use Vite +SWC to speed up the development process and reduce the amount of memory used by the frontend container (~200Mb instead of 1.5Gb before).